+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 * add citation format preview to telescope UI
+* add direct citation format selection in telescope UI with key mappings
 
 ### Bug Fixes